+++ R/plotSpectra2D.R
@@ -33,7 +33,7 @@
 #' One cannot remove frequencies from the interior of a 2D NMR data set and expect to get a meaningful
 #' contour plot, because doing so puts unrelated peaks adjacent in the data set.
 #' This would lead to contours being drawn that don't exist in the original data set.
-#' This function will check for missing frequencies and stop if any are found.
+#' This function will check for missing frequencies and stops if any are found.
 #'
 #' @section Scale:
 #' You can view the color scale for the plot via \code{\link{showScale}}.