+++ R/shiftSpectra2D.R
@@ -5,10 +5,10 @@
 #' or columns are thrown away and new rows or columns are introduced.  These new entries may
 #' be filled with zeros, or random noise.
 #' \itemize{
-#'   \item (+) x-shift - shift right: trim right, fill left
-#'   \item (-) x-shift - shift left: trim left, fill right
-#'   \item (+) y-shift - shift up: trim top, fill bottom
-#'   \item (-) y-shift - shift down: trim bottom, fill top
+#'   \item (+) \code{shiftF2} - shift right: trim right, fill left
+#'   \item (-) \code{shiftF2} - shift left: trim left, fill right
+#'   \item (+) \code{shiftF1} - shift up: trim top, fill bottom
+#'   \item (-) \code{shiftF1} - shift down: trim bottom, fill top
 #' }
 #'
 #' @param spectra An object of S3 class \code{\link{Spectra2D}}.
@@ -19,21 +19,38 @@
 #'
 #' @param shiftF1 As per \code{shiftF2}, but for the F1 dimension.
 #'
-#' @param fill Character.  One of \code{"zero"} or \code{"rnorm"} describing how to fill
-#'        the empty rows or columns that are created during shifting.  If \code{rnorm}
-#'        you may wish to set the seed.
-#'
+#' @param fill Aligning spectra requires that at least some spectra be shifted left/right
+#'        and up/down.  When a spectrum is shifted, spaces are opened that must be filled with something:
+#'  \itemize{
+#'    \item If \code{fill = "zeros"} the spaces are filled with zeros.
+#'    \item If \code{fill = "rnorm"} the spaces are filled with random noise.
+#'    \item If \code{fill = "noise"} the spaces are filled with an estimate of the noise from the
+#'      original spectrum.
+#'  }
 #' @return An object of S3 class \code{\link{Spectra2D}}.
 #'
 #' @author Bryan A. Hanson, DePauw University.
 #'
-#' @keywords multivariate
+#' @keywords utilities
 #'
 #' @export
+#' @examples
+#' data(MUD2)
+#' # Show the first two spectra, overlaid
+#' 
+#' mylvls <- seq(5, 35, 5)
+#' plotSpectra2D(MUD2, which = 1:2, lvls = list(mylvls, mylvls),
+#'   cols = list(rep("black", 7), rep("red", 7)),
+#'   main = "MUD2 Sample 1 (black) & Sample 2 (red)")
+#' 
+# Now shift Sample 2
+#' MUD2s <- shiftSpectra2D(MUD2, which = 2, shiftF1 = -2)
+#' plotSpectra2D(MUD2s, which = 1:2, lvls = list(mylvls, mylvls),
+#'   cols = list(rep("black", 7), rep("red", 7)),
+#'   main = "MUD2 Sample 1 (black) & Sample 2 (red)\n(samples aligned/overlap)")
 #'
-#'
-
-shiftSpectra2D <- function(spectra, which = NULL, shiftF2 = 0L, shiftF1 = 0L, fill = "zero") {
+#' 
+shiftSpectra2D <- function(spectra, which = NULL, shiftF2 = 0L, shiftF1 = 0L, fill = "noise") {
 
   .chkArgs(mode = 21L)
   chkSpectra(spectra)
@@ -43,13 +60,15 @@ shiftSpectra2D <- function(spectra, which = NULL, shiftF2 = 0L, shiftF1 = 0L, fi
   shiftF1 <- as.integer(shiftF1)
   if ((shiftF2 == 0L) & (shiftF1 == 0L)) stop("Both shiftF2 and shiftF1 are zero")
 
+  if (fill == "noise") NS <- .noiseSurface(spectra)
+  if (fill != "noise") NS <- NULL
+
   A <- .makeArray2(spectra, which)
   dimnames(A) <- NULL
-  A <- .shiftArray(A, xshift = shiftF2, yshift = shiftF1, fill = "zero")
+  A <- .shiftArray(A, xshift = shiftF2, yshift = shiftF1, fill = "zero", NS = NS)
   for (i in 1:length(which)) {
   	spectra$data[[ which[i] ]] <- A[i,,]
   }
-  spectra <- .trimZeros(spectra)
   
   chkSpectra(spectra)
   spectra