+++ tests/testthat.R
@@ -1,4 +1,0 @@
-# This file drives the testing
-library("testthat")
-library("ChemoSpec2D")
-test_check("ChemoSpec2D")